% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/met2model.LDNDC.R
\name{met2model.LDNDC}
\alias{met2model.LDNDC}
\title{Write LDNDC met files}
\usage{
met2model.LDNDC(
  in.path,
  in.prefix,
  outfolder,
  start_date,
  end_date,
  overwrite = FALSE,
  ...
)
}
\arguments{
\item{in.path}{path on disk where CF file lives}

\item{in.prefix}{prefix for each file}

\item{outfolder}{location where model specific output is written.}

\item{start_date}{start date of the results}

\item{end_date}{end date of the results}

\item{overwrite}{logical: replace output files if they already exist?}

\item{...}{Other arguments}
}
\value{
invisible(results)
}
\description{
Converts a met CF file to a model specific met file. The input
files are calld <in.path>/<in.prefix>.YYYY.cf
}
\author{
Henri Kajasilta
}
