% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.config.LDNDC.R
\name{write.config.LDNDC}
\alias{write.config.LDNDC}
\title{Write LDNDC configuration files}
\usage{
write.config.LDNDC(defaults, trait.values, settings, run.id)
}
\arguments{
\item{defaults}{list of defaults to process}

\item{trait.values}{vector of samples for a given trait}

\item{settings}{list of settings from pecan settings file}

\item{run.id}{id of run}
}
\value{
configuration file for LDNDC for given run
}
\description{
Writes a LDNDC config file.
}
\details{
Requires a pft xml object, a list of trait values for a single model run,
and the name of the file to create
}
\author{
Henri Kajasilta
}
